pyproject.toml
[tool.poetry]
name = "multi-tenancy-design-pattern-for-postgres"
version = "0.1.0"
description = ""
authors = ["amazingchow <jianzhou42@163.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.15.2"
email-validator = "^2.2.0"
fastapi = {extras = ["all"], version = "^0.115.12"}
psycopg = "^3.2.6"
psycopg-binary = "^3.2.6"
python-dotenv = "^1.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
uvicorn = "^0.34.2"

[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[tool.sort.config]
# More info: https://pypi.org/project/poetry-sort/
auto = true
case-sensitive = false
sort-python = false
format = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope="function"

[tool.isort]
# More info: https://pycqa.github.io/isort/docs/configuration/action_comments.html
profile = "hug"
src_paths = ["."]
known_local_folder = [
    "api",
    "core",
    "crud",
    "middlewares",
    "modles",
    "schemas",
]
sections = ["FUTURE", "STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]